% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Bootstrap.R
\name{BootEvents}
\alias{BootEvents}
\title{Bootstrap Events}
\usage{
BootEvents(y0, n0, y1, n1)
}
\arguments{
\item{y0}{Events per category in arm 0.}

\item{n0}{Subjects per category in arm 0.}

\item{y1}{Events per category in arm 1.}

\item{n1}{Subjects per category in arm 1.}
}
\value{
List containing the bootstrapped event counts.
}
\description{
Bootstrap the number of events observed in each cell of a contingency table,
keeping the number of subjects in each cell fixed.
}
