% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Rates.R
\name{MargRate}
\alias{MargRate}
\title{Marginal Rates}
\usage{
MargRate(y0, n0, y1, n1, weights = NULL)
}
\arguments{
\item{y0}{Events per category in arm 0.}

\item{n0}{Subjects per category in arm 0.}

\item{y1}{Events per category in arm 1.}

\item{n1}{Subjects per category in arm 1.}

\item{weights}{Stratum mixing weights.}
}
\value{
List containing:
\itemize{
  \item Marginal rates in each arm, `p0` and `p1`
  \item Per-stratum event rates, `r0` and `r1`.
  \item Stratum `weights`.
}
}
\description{
Calculates the marginal event rates in each arm, weighting by the total 
stratum size.
}
