% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Statistics.R
\name{RiskDiff}
\alias{RiskDiff}
\title{Asymptotic Risk Difference}
\usage{
RiskDiff(y0, n0, y1, n1, weights = NULL, alpha = 0.05)
}
\arguments{
\item{y0}{Events per category in arm 0.}

\item{n0}{Subjects per category in arm 0.}

\item{y1}{Events per category in arm 1.}

\item{n1}{Subjects per category in arm 1.}

\item{weights}{Stratum mixing weights.}

\item{alpha}{Type I error rate.}
}
\value{
Data.frame containing:
\itemize{
  \item 'Est', the estimated risk difference, arm 1 minus arm 0.
  \item The standard error 'SE'.
  \item The 'Lower' and 'Upper' confidence bounds.
  \item The asymptotic 'P' value.
}
}
\description{
Asymptotic Risk Difference
}
