% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Permutation.R
\name{TestNull}
\alias{TestNull}
\title{Test the Null Hypothesis via Permutation}
\usage{
TestNull(y0, n0, y1, n1, weights = NULL, alpha = 0.05, reps)
}
\arguments{
\item{y0}{Events per category in arm 0.}

\item{n0}{Subjects per category in arm 0.}

\item{y1}{Events per category in arm 1.}

\item{n1}{Subjects per category in arm 1.}

\item{weights}{Stratum mixing weights.}

\item{alpha}{Type I error rate.}

\item{reps}{Permutation replicates.}
}
\value{
Data.frame containing the marginal contrast estimates and permutation
  p-values for the risk difference, risk ratio, and odds ratio.
}
\description{
Test the Null Hypothesis via Permutation
}
